Adds tests for emailing.
Tests the emailing functinality by testing:

- The addresses the email is sent to
- The subject

To make these tests not actually send emails, we inject a mocked version of our emailer. Currently, the class used by `@sendgrid/mail` is not exported. Fortunately, by testing `index` which has a default which is the class, the compiler will throw an error if we pass in something that doesn't match. Comments were added in the test by our mock and in the code itself when we have to accep an `any`. Hopefully, we can import the type at some point which can allow us to be more specific with out types.