Fixes tests to properly check `to` field.
In f9a8e610120ba4fcfb406db540d646a163569c01 we fixed sending emails to
multiple recipients to [work like sendgrid/mail expects it
to](https://github.com/sendgrid/sendgrid-nodejs/blob/master/use-cases/single-email-multiple-recipients.md).

It seems like when writing the tests, this change wasn't merged in so
the tests were being written to work in the old, wrong way (`to` as a
comma-separated string) instead of the new, correct way (`to` as an
array of strings).

This commit updates the tests to work with the new way.