Updates `weekly-summary-typescript` dep to 1.0.
The `weekly-summary-typescript` package is what does all of the work for
retrieving the recently closed pull requests. It's [1.0.0
release](https://github.com/tmr08c/weekly-summary-typescript/releases/tag/v1.0.0)
has a breaking change where you not have to explicitly pass in the
GitHub auth token, instead of needing to have it set as a specific
environment  variable. Fortunately, this change was pretty
straightforward for this project.